+++ code/core/src/main/java/org/betonquest/betonquest/quest/action/chest/ChestTakeActionFactory.java
@@ -1,7 +1,9 @@
 package org.betonquest.betonquest.quest.action.chest;
 
 import org.betonquest.betonquest.api.QuestException;
+import org.betonquest.betonquest.api.identifier.ActionIdentifier;
 import org.betonquest.betonquest.api.instruction.Argument;
+import org.betonquest.betonquest.api.instruction.FlagArgument;
 import org.betonquest.betonquest.api.instruction.Instruction;
 import org.betonquest.betonquest.api.instruction.type.ItemWrapper;
 import org.betonquest.betonquest.api.quest.action.NullableActionAdapter;
@@ -9,6 +11,7 @@ import org.betonquest.betonquest.api.quest.action.PlayerAction;
 import org.betonquest.betonquest.api.quest.action.PlayerActionFactory;
 import org.betonquest.betonquest.api.quest.action.PlayerlessAction;
 import org.betonquest.betonquest.api.quest.action.PlayerlessActionFactory;
+import org.betonquest.betonquest.api.service.action.ActionManager;
 import org.bukkit.Location;
 
 import java.util.List;
@@ -19,9 +22,17 @@ import java.util.List;
 public class ChestTakeActionFactory implements PlayerActionFactory, PlayerlessActionFactory {
 
     /**
-     * Create the chest take action factory.
+     * The action manager.
      */
-    public ChestTakeActionFactory() {
+    private final ActionManager actionManager;
+
+    /**
+     * Create the chest take action factory.
+     *
+     * @param actionManager the action manager
+     */
+    public ChestTakeActionFactory(final ActionManager actionManager) {
+        this.actionManager = actionManager;
     }
 
     @Override
@@ -37,6 +48,9 @@ public class ChestTakeActionFactory implements PlayerActionFactory, PlayerlessAc
     private NullableActionAdapter createChestTakeAction(final Instruction instruction) throws QuestException {
         final Argument<Location> location = instruction.location().get();
         final Argument<List<ItemWrapper>> item = instruction.item().list().get();
-        return new NullableActionAdapter(new ChestTakeAction(location, item));
+        final FlagArgument<Boolean> abort = instruction.bool().getFlag("abort", true);
+        final Argument<List<ActionIdentifier>> failActions = instruction.identifier(ActionIdentifier.class)
+                .list().get("fail").orElse(null);
+        return new NullableActionAdapter(new ChestTakeAction(location, actionManager, item, abort, failActions));
     }
 }