+++ code/core/src/main/java/org/betonquest/betonquest/quest/action/take/TakeActionFactory.java
@@ -2,7 +2,9 @@ package org.betonquest.betonquest.quest.action.take;
 
 import org.betonquest.betonquest.api.QuestException;
 import org.betonquest.betonquest.api.config.Localizations;
+import org.betonquest.betonquest.api.identifier.ActionIdentifier;
 import org.betonquest.betonquest.api.instruction.Argument;
+import org.betonquest.betonquest.api.instruction.FlagArgument;
 import org.betonquest.betonquest.api.instruction.Instruction;
 import org.betonquest.betonquest.api.instruction.type.ItemWrapper;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
@@ -10,6 +12,7 @@ import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.api.quest.action.OnlineActionAdapter;
 import org.betonquest.betonquest.api.quest.action.PlayerAction;
 import org.betonquest.betonquest.api.quest.action.PlayerActionFactory;
+import org.betonquest.betonquest.api.service.action.ActionManager;
 import org.betonquest.betonquest.data.PlayerDataStorage;
 import org.betonquest.betonquest.quest.action.IngameNotificationSender;
 import org.betonquest.betonquest.quest.action.NoNotificationSender;
@@ -39,17 +42,24 @@ public class TakeActionFactory implements PlayerActionFactory {
     private final Localizations localizations;
 
     /**
+     * The action manager.
+     */
+    private final ActionManager actionManager;
+
+    /**
      * Create the take action factory.
      *
      * @param loggerFactory     the logger factory to create a logger for the actions
      * @param playerDataStorage the storage providing player data
      * @param localizations     the {@link Localizations} instance
+     * @param actionManager     the action manager
      */
     public TakeActionFactory(final BetonQuestLoggerFactory loggerFactory, final PlayerDataStorage playerDataStorage,
-                             final Localizations localizations) {
+                             final Localizations localizations, final ActionManager actionManager) {
         this.loggerFactory = loggerFactory;
         this.playerDataStorage = playerDataStorage;
         this.localizations = localizations;
+        this.actionManager = actionManager;
     }
 
     @Override
@@ -58,7 +68,11 @@ public class TakeActionFactory implements PlayerActionFactory {
         final Argument<List<CheckType>> checkOrder = getCheckOrder(instruction);
         final Argument<List<ItemWrapper>> questItems = instruction.item().list().get();
         final NotificationSender notificationSender = getNotificationSender(instruction, log);
-        return new OnlineActionAdapter(new TakeAction(playerDataStorage, notificationSender, questItems, checkOrder));
+        final FlagArgument<Boolean> abort = instruction.bool().getFlag("abort", true);
+        final Argument<List<ActionIdentifier>> failActions = instruction.identifier(ActionIdentifier.class)
+                .list().get("fail").orElse(null);
+        return new OnlineActionAdapter(new TakeAction(playerDataStorage, notificationSender, actionManager,
+                questItems, checkOrder, abort, failActions));
     }
 
     /**