+++ CHANGELOG.md
@@ -71,6 +71,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `conversation.block_item_transfer` option with per-conversation override to prevent item moving while in a conversation
 - new notification category `conversation_nothing_to_start`
 - `check` flag to `menu` action to respect the menu's open conditions
+- `abort` flag and `fail` actions to (chest) take action
 ### Changed
 - Spigot is no longer supported, paper is now required 
 - message.yml file was deleted and instead the lang folder now contains all translations